Make deref'ing Gc "infallible" by default.
I say "infallible" because it is allowed to panic, as all safe code is.

This improves the usage ergonomics considerably in the case where code
can reasonably expect that "stored" references are either:
- strongly held, or
- not interrupted by a collection barrier.
If these invariants aren't true, the old behavior is available as
Gc::try_as_ref and Gc::try_as_mut, which truly are _infallible_, but
return an Option (much like rc::Weak).

Because of Deref and DerefMut implementations, Gc "methods" are now
associated functions to avoid conflicts with the target type.